Move remaining notification logic to frontend (#1313)
* Use source name in route (not UUID)

* Fetch runtime instanceID in page load

* Fix comparison of Go and Node source names

* Use source names not IDs in sidebar links

* Create YAML artifact when submitting source form

* Use source name as identifier in `WorkspaceContainer`

* Remove unused prop

* Use new Rename API

* Use new Delete API

* Generate client-side code

* Use model name in routes (not UUID)

* Generate client-side code

* Use new Rename API

* Use new Delete API

* Create a model for a given source using `PutAndMigrate`

* Create a new blank model with `PutFile`

* Add util function

* Update model with `PutAndMigrate`

* Various cleanups

* Fixing rename and delete

* Upload drag-and-dropped files w/ `PutAndMigrate`

* Use `GetFile` in `+page.ts`

* Update url after renaming a source

* Add defensive checks for source id

* Clean up `SourceMenuItems.svelte`

* Move to next source after a delete

* Use `ListFiles` not `ListCatalogObjects`

* Generate client-side code

* Update runtime version

* Add check for a valid name in `SourceWorkspaceHeader`

* Update `+page.ts` to use `GetFile`

* Cleanup `ModelBody.svelte`

* Fix merge

* In sidebar, use `ListFiles` and `PutFileAndMigrate`

* Fix deleting a model

* Use model name from URL, not active entity ID

* Update URL after renaming a model

* Update URL after renaming a model (from modal)

* Key on `modelName` not ID

* Remove old code

* Add yaml serializer

* Use service catalog

* Add yamlInternalRepresentation class

* Add class methods

* Mutate document directly

* Dashboard names are dynamic

* Refactor metric components

* Use source name in route (not UUID)

* Fetch runtime instanceID in page load

* Fix comparison of Go and Node source names

* Use source names not IDs in sidebar links

* Create YAML artifact when submitting source form

* Use source name as identifier in `WorkspaceContainer`

* Remove unused prop

* Use new Rename API

* Use new Delete API

* Generate client-side code

* Fixing rename and delete

* Upload drag-and-dropped files w/ `PutAndMigrate`

* Use `GetFile` in `+page.ts`

* Update url after renaming a source

* Add defensive checks for source id

* Clean up `SourceMenuItems.svelte`

* Move to next source after a delete

* Use `ListFiles` not `ListCatalogObjects`

* Generate client-side code

* Update runtime version

* Add check for a valid name in `SourceWorkspaceHeader`

* Fixing local files

* Adding common code

* Add delete menu option

* Rename store

* Fixing build

* Fixing configs

* Fixing rename and updating metrics assets

* Saving and returning schema from catalog service (#1283)

* Mirror models to nodejs (#1284)

* Fix model create,rename and delete

* Adding sync with nodejs states

* Adding framework for UI tests (#1291)

* Adding a pipeline for errors from migrate calls (#1295)

* Fixing delete not propagating (#1302)

* Metrics migration (#1282)

* Serialize json blocks as YAML

* Yaml fixes

* Update model schema field extraction and duckdb types

* Add quickstart button

* Move store and create API for quick metrics

* Remove repoID from queries

* Adding partailly valid metrics to catalog

* Add autogenerate from model (#1308)

* Fix reactivity bug (#1311)

* Address Browserlist `caniuse` dependency warning

* Add frontend notifications for CRUD operations

* Make imports cleaner for notifications

* Remove notifications-over-websocket

* Delete old file

* Delete more old files

* Fix imports

* Fix lint

* Remove old comment

* Fix lint
