+++ web-local/src/lib/components/metrics-definition/MetricsDefinitionExploreMetricsButton.svelte
@@ -5,11 +5,11 @@
     MetricsEventScreenName,
     MetricsEventSpace,
   } from "@rilldata/web-local/lib/metrics/service/MetricsTypes";
+  import { navigationEvent } from "../../metrics/initMetrics";
   import { Button } from "../button";
   import ExploreIcon from "../icons/Explore.svelte";
   import Tooltip from "../tooltip/Tooltip.svelte";
   import TooltipContent from "../tooltip/TooltipContent.svelte";
-  import { navigationEvent } from "../../metrics/initMetrics";
 
   export let metricsInternalRep;
   export let metricsDefName;
