+++ web-local/src/lib/components/navigation/dashboards/MetricsDefinitionAssets.svelte
@@ -19,8 +19,8 @@
   import { LIST_SLIDE_DURATION } from "@rilldata/web-local/lib/application-config";
   import { runtimeStore } from "@rilldata/web-local/lib/application-state-stores/application-store";
   import {
-    fileArtifactsStore,
     FileArtifactsData,
+    fileArtifactsStore,
   } from "@rilldata/web-local/lib/application-state-stores/file-artifacts-store.js";
   import { metricsTemplate } from "@rilldata/web-local/lib/application-state-stores/metrics-internal-store";
   import Model from "@rilldata/web-local/lib/components/icons/Model.svelte";
@@ -33,14 +33,14 @@
   import { navigationEvent } from "../../../metrics/initMetrics";
   import { queryClient } from "../../../svelte-query/globalQueryClient";
   import Cancel from "../../icons/Cancel.svelte";
+  import EditIcon from "../../icons/EditIcon.svelte";
   import { default as Explore } from "../../icons/Explore.svelte";
+  import MetricsIcon from "../../icons/Metrics.svelte";
   import { MenuItem } from "../../menu";
   import MetricsDefinitionSummary from "../../metrics-definition/MetricsDefinitionSummary.svelte";
   import NavigationEntry from "../NavigationEntry.svelte";
   import NavigationHeader from "../NavigationHeader.svelte";
   import RenameAssetModal from "../RenameAssetModal.svelte";
-  import MetricsIcon from "../../icons/Metrics.svelte";
-  import EditIcon from "../../icons/EditIcon.svelte";
 
   $: instanceId = $runtimeStore.instanceId;
 
@@ -164,7 +164,7 @@
           )}
           {@const hasSourceError =
             selectionError !== SourceModelValidationStatus.OK &&
-            selectionError !== ""}
+            selectionError !== ""} -->
           <MenuItem
             icon
             disabled={hasSourceError}
