+++ web-local/src/lib/components/navigation/sources/SourceMenuItems.svelte
@@ -18,7 +18,6 @@
     DerivedTableStore,
     PersistentTableStore,
   } from "@rilldata/web-local/lib/application-state-stores/table-stores";
-  import { getFileFromName } from "@rilldata/web-local/lib/util/entity-mappers";
   import { createModelFromSource } from "@rilldata/web-local/lib/components/navigation/models/createModel";
   import { autoCreateMetricsDefinitionForSource } from "@rilldata/web-local/lib/redux-store/source/source-apis";
   import { derivedProfileEntityHasTimestampColumn } from "@rilldata/web-local/lib/redux-store/source/source-selectors";
@@ -28,6 +27,7 @@
     useSourceFromYaml,
     useSourceNames,
   } from "@rilldata/web-local/lib/svelte-query/sources";
+  import { getFileFromName } from "@rilldata/web-local/lib/util/entity-mappers";
   import { createEventDispatcher, getContext } from "svelte";
   import { EntityType } from "../../../../common/data-modeler-state-service/entity-state-service/EntityStateService";
   import {
