+++ web-local/src/lib/components/navigation/sources/createSource.ts
@@ -7,9 +7,10 @@ import {
 import { EntityType } from "@rilldata/web-local/common/data-modeler-state-service/entity-state-service/EntityStateService";
 import { dataModelerService } from "@rilldata/web-local/lib/application-state-stores/application-store";
 import { fileArtifactsStore } from "@rilldata/web-local/lib/application-state-stores/file-artifacts-store";
-import { getFileFromName } from "@rilldata/web-local/lib/util/entity-mappers";
 import { queryClient } from "@rilldata/web-local/lib/svelte-query/globalQueryClient";
 import type { UseMutationResult } from "@sveltestack/svelte-query";
+import { getFileFromName } from "../../../util/entity-mappers";
+import { notifications } from "../../notifications";
 
 export async function createSource(
   instanceId: string,
@@ -37,5 +38,6 @@ export async function createSource(
   await queryClient.invalidateQueries(
     getRuntimeServiceListFilesQueryKey(instanceId)
   );
+  notifications.send({ message: `Created source ${tableName}` });
   return [];
 }
