+++ web-local/src/lib/components/workspace/explore/time-controls/TimeRangeNameSelector.svelte
@@ -1,5 +1,4 @@
 <script lang="ts">
-  import { runtimeStore } from "@rilldata/web-local/lib/application-state-stores/application-store";
   import type {
     TimeRangeName,
     TimeSeriesTimeRange,
@@ -9,7 +8,6 @@
     MetricsExplorerEntity,
     metricsExplorerStore,
   } from "../../../../application-state-stores/explorer-stores";
-  import { useMetaQuery } from "../../../../svelte-query/queries/metrics-views/metadata";
   import { onClickOutside } from "../../../../util/on-click-outside";
   import { FloatingElement } from "../../../floating-element";
   import Calendar from "../../../icons/Calendar.svelte";
