+++ web-local/src/lib/components/workspace/source/SourceWorkspaceHeader.svelte
@@ -22,7 +22,7 @@
   import Import from "../../icons/Import.svelte";
   import RefreshIcon from "../../icons/RefreshIcon.svelte";
   import Source from "../../icons/Source.svelte";
-  import notifications from "../../notifications";
+  import { notifications } from "../../notifications";
   import Tooltip from "../../tooltip/Tooltip.svelte";
   import TooltipContent from "../../tooltip/TooltipContent.svelte";
   import WorkspaceHeader from "../core/WorkspaceHeader.svelte";
