+++ web-local/src/routes/(application)/dashboard/[name]/edit/+page.svelte
@@ -2,8 +2,8 @@
   import { useRuntimeServiceGetFile } from "@rilldata/web-common/runtime-client";
   import { EntityType } from "@rilldata/web-local/common/data-modeler-state-service/entity-state-service/EntityStateService";
   import { runtimeStore } from "@rilldata/web-local/lib/application-state-stores/application-store";
-  import { getFileFromName } from "@rilldata/web-local/lib/util/entity-mappers";
   import { MetricsDefinitionWorkspace } from "@rilldata/web-local/lib/components/workspace";
+  import { getFileFromName } from "@rilldata/web-local/lib/util/entity-mappers";
 
   export let data;
